requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
WTForms==3.1.1
email-validator==2.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
Unidecode==1.3.8
pytest==7.4.3
pytest-cov==4.1.0
email-validator==2.1.0
Flask==3.0.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
Unidecode==1.3.8
Werkzeug==3.0.1
WTForms==3.1.1
